create component for onboading
